Added Crouch toggle and stealth tree function to player controller
Added function and variables for stealth to track isMoving, isCrouched and if player is hiding in environment. bas
Added crouchModifier, stopModifier and envModifier to add to baseStealthStat variables
baseStealthStat set to 10.... all modifiers add 5 to the base stat.
modifiers will be added if player is crouching or not ( ie. base not crouching + not moving  = 10 ||base crouching and not moving = 15)