XP System Activated
added XP Manager to Pigeon BP, added functionality to Pigeon BP in Begin Play to create a list of components that use XP Interface, and now uses that list to communicate level ups between xp manager and all other components that use the xp interface. This is also modular so any new components added that implement the xp interface can be notified of level up events.